# Consolidate multiple tensor copies to reduce API overhead