@slaren 2025-09-21
This cannot be merged in this way, it's too hacky and not guaranteed to work every buffer type or in more complex scenarios, such as multiple GPUs. I am also currently making some significant changes to `ggml_backend_sched`, and would prefer to avoid introducing significant conflicts.